backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
python-multipart>=0.0.6
pandas>=2.0.0
openpyxl>=3.1.0
reportlab>=4.0.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
